+++ zingolib/src/lightclient/send.rs
@@ -119,7 +119,7 @@ fn resolve_transmit_route(
 ) -> Result<Option<std::net::SocketAddr>, LightClientError> {
     use crate::mixnet::{MixnetNotReady, MixnetRoute};
     match (has_indexer, route) {
-        (_, Ok(MixnetRoute::Mixnet(tunnel))) => Ok(Some(tunnel.into_addr())),
+        (_, Ok(MixnetRoute::Mixnet(conduit))) => Ok(Some(conduit.socks5())),
         (true, Ok(MixnetRoute::Clearnet)) => Ok(None),
         (false, Ok(MixnetRoute::Clearnet)) => Err(LightClientError::Offline),
         (false, Err(MixnetNotReady::Unattached)) => Err(LightClientError::Offline),